Input system reorganized to fix several coroutine bugs
The input system itself no longer keeps track of nested key sequences.  However, prefix commands are still permitted to use the minibuffer, and in this way nested key sequences can still occur.

call_interactively is no longer a coroutine.  Instead, use run_interactively if you want to wait for the command to finish and handle any errors it throws.  If you want to run a command that actually interacts with the input system, use input_run_command instead, but be sure to call it with an interactive_context created by input_make_interactive_context.