Finish correcting replacements of `for each...in' loops
1. contrib/xrev/content/xrev.js: key_event_props is an array and we're
interested in the elements, not indices of the array.

2. modules/user-agent-policy.jsx: policies is not iterable.

3. modules/utils.js: for each...in loops are forgiving when the
iterable object is undefined (they just logically skip the loop).
Unfortunately, for...of loops aren't as forgiving.  This was the cause
for the lack of completion of search results in the minibuffer.

I still occasionally get a:

reference to undefined property "count"" {file:
"chrome://conkeror/content/completers.js" line: 396}

error, which I think didn't occur previously (0864afd1b830d), but I
can't pinpoint the reason.  It doesn't seem to affect any
functionality, though.

All 47 replacements of the `for each...in' loops have been inspected.