Remove atom_service completely
Fix the `atom_service is not defined' error and the lack of bold for
the completion descriptions.

Technically, nsIAtomService is actually still provided by Waterfox
2019-12, so one could just uncomment the line defining atom_service,
instead.

However, getting (and setting) the props in getCellProperties is
obsolete starting from Gecko 22, according to:
https://developer.mozilla.org/en-US/docs/Mozilla/Tech/XPCOM/Reference/Interface/nsITreeView
and for the purposes of returning the cell properties, a simple string
will suffice.

I haven't tested on Palemoon, but AFAIR the divergence point for its
most recent version is after Firefox 22, so it shouldn't be adversely
affected.